Widget Refactoring
Now each widget must implement the Dancer2::Plugin::LiteBlog::Widget
interface.

Namely, it has a 'root' attribute which defaults to the appdir.
It has to implement the 'elements' method, which is called by the
engine in the before_template hook, to populate the widget.

This must return a list of objects that the template will iterate over.