Widgets can now declare their own routes
The old Routes.pm is deprecated. Now, each Widget is responsible for
implementing their own routes, via the keyword 'declare_route'.
If they do so, they should implement has_routes as a sub that returns 1.

Working example in this commit with the Blog widget and its 'permalink'
route that can be mounted to any prefix with the 'mount' setting.