First working version of the Splash widget
Able to display a big background image on the home page.
On top of it, centered, a big Title, a baseline and a call-to-action
button.

A great widget for site that wants to drive their visitors to their
main content right away. Can be useful for collecting subscribers to
a newsletter or driving users to a featured content.

When enabled, this widget disables the Hero header of the home page.
Maybe the Hero banner should be a widget itself?